pyproject.toml
[tool.poetry]
name = "python-devbooster"
version = "0.1.0"
description = ""
authors = ["vvpreo <vvpreo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.2"
pyyaml = "^6.0.1"
psutil = "^5.9.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
